Codechange: Split native and emscripten bootstrap windows.
This simplifies testing defines and makes each part self-contained.
